feat(exception) adding global exceptions

+++ src/main/java/com/apisql/ApiSQL/service/CaminhaoService.java
@@ -2,6 +2,7 @@ package com.apisql.ApiSQL.service;
 
 import com.apisql.ApiSQL.dto.CaminhaoRequestDTO;
 import com.apisql.ApiSQL.dto.CaminhaoResponseDTO;
+import com.apisql.ApiSQL.exception.ResourceNotFoundException;
 import com.apisql.ApiSQL.model.Caminhao;
 import com.apisql.ApiSQL.model.Segmento;
 import com.apisql.ApiSQL.model.Unidade;
@@ -9,12 +10,12 @@ import com.apisql.ApiSQL.repository.CaminhaoRepository;
 import com.apisql.ApiSQL.repository.SegmentoRepository;
 import com.apisql.ApiSQL.repository.UnidadeRepository;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import jakarta.persistence.EntityNotFoundException;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -26,19 +27,18 @@ public class CaminhaoService {
     private final UnidadeRepository unidadeRepository;
     private final ObjectMapper objectMapper;
 
-    public CaminhaoService(CaminhaoRepository caminhaoRepository, ObjectMapper objectMapper, SegmentoRepository segmentoRepository, UnidadeRepository unidadeRepository) {
-
+    public CaminhaoService(CaminhaoRepository caminhaoRepository, SegmentoRepository segmentoRepository, UnidadeRepository unidadeRepository, ObjectMapper objectMapper) {
         this.caminhaoRepository = caminhaoRepository;
-        this.objectMapper = objectMapper;
         this.segmentoRepository = segmentoRepository;
         this.unidadeRepository = unidadeRepository;
+        this.objectMapper = objectMapper;
     }
 
     public List<CaminhaoResponseDTO> findAll() {
         List<Caminhao> caminhoes = caminhaoRepository.findAll();
-        return caminhoes.stream().map(
-                c -> objectMapper.convertValue(c, CaminhaoResponseDTO.class)
-        ).toList();
+        return caminhoes.stream()
+                .map(c -> objectMapper.convertValue(c, CaminhaoResponseDTO.class))
+                .toList();
     }
 
     public CaminhaoResponseDTO findById(Integer id) {
@@ -46,42 +46,57 @@ public class CaminhaoService {
         if (response.isPresent()) {
             return objectMapper.convertValue(response.get(), CaminhaoResponseDTO.class);
         }
-        throw new EntityNotFoundException("Caminhao nao encontrado");
+        throw new ResourceNotFoundException("Caminhão não encontrado com ID: " + id);
     }
 
     @Transactional
     public CaminhaoResponseDTO save(CaminhaoRequestDTO dto) {
-        Caminhao caminhao = new Caminhao();
         Segmento segmento = segmentoRepository.findById(dto.getIdSegmento())
-                .orElseThrow(() -> new EntityNotFoundException("Segmento não encontrado com ID: " + dto.getIdSegmento()));
-
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID Segmento inválido."));
         Unidade unidade = unidadeRepository.findById(dto.getIdUnidade())
-                .orElseThrow(() -> new EntityNotFoundException("Unidade não encontrada com ID: " + dto.getIdUnidade()));
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID Unidade inválido."));
 
+        Caminhao caminhao = new Caminhao();
         caminhao.setChassi(dto.getChassi());
-        caminhao.setSegmento(segmento);
-        caminhao.setUnidade(unidade);
         caminhao.setPlaca(dto.getPlaca());
         caminhao.setModelo(dto.getModelo());
         caminhao.setAnoFabricacao(dto.getAnoFabricacao());
         caminhao.setNumeroFrota(dto.getNumeroFrota());
+        caminhao.setSegmento(segmento);
+        caminhao.setUnidade(unidade);
+
         Caminhao savedCaminhao = caminhaoRepository.save(caminhao);
         return objectMapper.convertValue(savedCaminhao, CaminhaoResponseDTO.class);
     }
 
     @Transactional
-    public CaminhaoResponseDTO update(Integer id, Caminhao caminhaoAtualizado) {
+    public CaminhaoResponseDTO update(Integer id, CaminhaoRequestDTO dto) {
         Caminhao caminhao = caminhaoRepository.findById(id)
-                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "Caminhao com id:" + id + " não encontrado"));
+                .orElseThrow(() -> new ResourceNotFoundException("Caminhão com id:" + id + " não encontrado para atualização"));
 
-        caminhao.setChassi(caminhaoAtualizado.getChassi());
-        return objectMapper.convertValue(caminhaoRepository.save(caminhaoAtualizado), CaminhaoResponseDTO.class);
+        Segmento segmento = segmentoRepository.findById(dto.getIdSegmento())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID Segmento inválido."));
+        Unidade unidade = unidadeRepository.findById(dto.getIdUnidade())
+                .orElseThrow(() -> new ResponseStatusException(HttpStatus.BAD_REQUEST, "ID Unidade inválido."));
+
+        caminhao.setChassi(dto.getChassi());
+        caminhao.setPlaca(dto.getPlaca());
+        caminhao.setModelo(dto.getModelo());
+        caminhao.setAnoFabricacao(dto.getAnoFabricacao());
+        caminhao.setNumeroFrota(dto.getNumeroFrota());
+        caminhao.setSegmento(segmento);
+        caminhao.setUnidade(unidade);
+
+        caminhao.setUpdatedAt(LocalDateTime.now());
+        Caminhao updatedCaminhao = caminhaoRepository.save(caminhao);
+        return objectMapper.convertValue(updatedCaminhao, CaminhaoResponseDTO.class);
     }
 
-
+    @Transactional
     public void deleteById(Integer id) {
+        if (!caminhaoRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Caminhão com id:" + id + " não encontrado para exclusão");
+        }
         caminhaoRepository.deleteById(id);
     }
-
-
 }
